requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
black==23.7.0
certifi==2023.7.22
cfgv==3.4.0
charset-normalizer==3.2.0
click==8.1.7
distlib==0.3.7
Django==4.2.4
djangorestframework==3.14.0
filelock==3.12.2
flake8==6.1.0
frozenlist==1.4.0
identify==2.5.27
idna==3.4
isort==5.12.0
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
nodeenv==1.8.0
openai==0.28.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pre-commit==3.3.3
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pyflakes==3.1.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
soupsieve==2.4.1
sqlparse==0.4.4
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.4
virtualenv==20.24.3
yarl==1.9.2
